Extract generator and watcher functions
